+++ vocode/streaming/telephony/conversation/zoom_dial_in.py
@@ -1,18 +1,13 @@
 import logging
 from typing import Optional
-from twilio.rest import Client
-from vocode.streaming.agent.base_agent import BaseAgent
 from vocode.streaming.models.agent import AgentConfig
 from vocode.streaming.models.synthesizer import SynthesizerConfig
-from vocode.streaming.models.telephony import CallConfig, TwilioConfig
+from vocode.streaming.models.telephony import TwilioConfig
 from vocode.streaming.models.transcriber import TranscriberConfig
-from vocode.streaming.synthesizer.base_synthesizer import BaseSynthesizer
 from vocode.streaming.telephony.config_manager.base_config_manager import (
     BaseConfigManager,
 )
 from vocode.streaming.telephony.conversation.outbound_call import OutboundCall
-from vocode.streaming.transcriber.base_transcriber import BaseTranscriber
-from vocode.streaming.utils import create_conversation_id
 
 
 class ZoomDialIn(OutboundCall):
@@ -25,12 +20,16 @@ class ZoomDialIn(OutboundCall):
         from_phone: str,
         config_manager: BaseConfigManager,
         agent_config: AgentConfig,
-        transcriber_config: TranscriberConfig,
-        synthesizer_config: SynthesizerConfig,
+        transcriber_config: Optional[TranscriberConfig] = None,
+        synthesizer_config: Optional[SynthesizerConfig] = None,
         twilio_config: Optional[TwilioConfig] = None,
         conversation_id: Optional[str] = None,
         logger: Optional[logging.Logger] = None,
     ):
+        digits = f"wwww{zoom_meeting_id}#"
+        if zoom_meeting_password:
+            digits += f"wwwwwwww*{zoom_meeting_password}#"
+
         super().__init__(
             base_url=base_url,
             to_phone=zoom_number,
@@ -42,34 +41,10 @@ class ZoomDialIn(OutboundCall):
             twilio_config=twilio_config,
             conversation_id=conversation_id,
             logger=logger,
+            mobile_only=False,
+            digits=digits,
         )
+
         self.zoom_number = zoom_number
         self.zoom_meeting_id = zoom_meeting_id
         self.zoom_meeting_password = zoom_meeting_password
-        self.from_phone = from_phone
-
-    def start(self):
-        self.validate_outbound_call(
-            self.zoom_number,
-            self.from_phone,
-            mobile_only=False,
-        )
-        digits = f"ww{self.zoom_meeting_id}#"
-        if self.zoom_meeting_password:
-            digits += f"wwww*{self.zoom_meeting_password}#"
-        self.logger.debug("Sending digits %s to the call", digits)
-        twilio_sid = self.create_twilio_call(
-            self.zoom_number,
-            self.from_phone,
-            digits=digits,
-        )
-        call_config = CallConfig(
-            transcriber_config=self.transcriber_config,
-            agent_config=self.agent_config,
-            synthesizer_config=self.synthesizer_config,
-            twilio_config=self.twilio_config,
-            twilio_sid=twilio_sid,
-            twilio_from=self.from_phone,
-            twilio_to=digits,
-        )
-        self.config_manager.save_config(self.conversation_id, call_config)
